pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "authproviderconnector"
version = "0.1.0"
description = "Auth0-based authentication connector service"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Mitou", email = "dev@example.com" }]
dependencies = [
  "fastapi>=0.115.2,<0.116",
  "uvicorn[standard]>=0.30.3,<0.31",
  "sqlalchemy>=2.0.32,<3.0",
  "asyncpg>=0.29.0,<0.30",
  "authlib>=1.3.1,<1.4",
  "pyjwt[crypto]>=2.9.0,<3.0",
  "pydantic-settings>=2.4.0,<3.0",
  "alembic>=1.13.3,<2.0",
  "itsdangerous>=2.2.0,<3.0",
  "psycopg2-binary>=2.9.0,<3.0",
  "httpx>=0.27.2,<0.28",
]

[dependency-groups]
dev = [
  "ruff>=0.7.0,<0.8",
  "mypy>=1.11.2,<2.0",
  "pytest>=8.3.2,<9.0",
  "pytest-asyncio>=0.23.8,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
  "authproviderconnector",
]

[tool.setuptools.package-data]
authproviderconnector = ["py.typed"]

[project.scripts]
up = "cli:run"

[tool.uv]
managed = true
default-groups = ["dev"]
